Add authorization.

+++ controllers/feed.js
@@ -1,4 +1,5 @@
 const Post = require('../models/post');
+const User = require('../models/user');
 const {
   throwErrorIfInvalid,
   handleAsyncError,
@@ -39,21 +40,32 @@ exports.postPosts = (req, res, next) => {
 
   const url = req.file.path.replace('\\', '/');
   const { title, content } = req.body;
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: url,
-    creator: {
-      name: 'Name'
-    }
+    creator: req.userId
   });
 
   post
     .save()
     .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
         message: 'Post created successfully!',
-        post: result
+        post: post,
+        creator: {
+          _id: creator._id,
+          name: creator.name
+        }
       });
     })
     .catch(err => handleAsyncError(err, next));
@@ -99,6 +111,13 @@ exports.putPost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if (result.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized.');
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== result.imageUrl) {
         clearImage(result.imageUrl);
       }
@@ -125,14 +144,59 @@ exports.deletePost = (req, res, next) => {
         throw error;
       }
 
-      //Check logged in user
+      if (result.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized.');
+        error.statusCode = 403;
+        throw error;
+      }
 
       clearImage(result.imageUrl);
 
       return Post.findByIdAndRemove(postId);
     })
     .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       res.status(200).json({ message: 'Deleted message' });
+    })
+    .catch(err => handleAsyncError(err, next));
+};
+
+exports.getStatus = (req, res, next) => {
+  User.findById(req.userId)
+    .then(user => {
+      if (!user) {
+        const error = new Error('Could not find post.');
+        error.statusCode = 404;
+        throw error;
+      }
+
+      return res.status(200).json({ status: user.status });
+    })
+    .catch(err => handleAsyncError(err, next));
+};
+
+exports.patchStatus = (req, res, next) => {
+  const newStatus = req.body.status;
+
+  User.findById(req.userId)
+    .then(user => {
+      if (!user) {
+        const error = new Error('Could not find post.');
+        error.statusCode = 404;
+        throw error;
+      }
+
+      user.status = newStatus;
+      return user.save();
+    })
+    .then(result => {
+      return res.status(200).json({ status: newStatus });
     })
     .catch(err => handleAsyncError(err, next));
 };
